# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: string;
  name: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  eventDate: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  dataDeleteDate: string;
  attendeeCount: number;
  checkedInCount: number;
  checkInUrl: string;
}

export interface EventInput {
  /** @minLength 1 */
  name: string;
  organizerEmail: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  eventDate: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  dataDeleteDate: string;
}

export interface AttendeeImportRow {
  name?: string;
  email: string;
  /** @minLength 1 */
  couponCode: string;
}

export interface AttendeeImportInput {
  /** @minItems 1 */
  attendees: AttendeeImportRow[];
}

export interface ImportResult {
  imported: number;
  rejected: number;
  errors?: string[];
}

export interface EventSummary {
  total: number;
  checkedIn: number;
  remaining: number;
  emailSent: number;
}

export interface FinalReportResult {
  sent: boolean;
  total: number;
  checkedIn: number;
  didNotShow: number;
}

export interface AttendeeStatus {
  name?: string;
  email: string;
  checkedIn: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  emailSent: boolean;
}

export interface CheckInInput {
  email: string;
}

export type CheckInResultStatus = typeof CheckInResultStatus[keyof typeof CheckInResultStatus];


export const CheckInResultStatus = {
  checked_in: 'checked_in',
  already_checked_in: 'already_checked_in',
} as const;

export interface CheckInResult {
  status: CheckInResultStatus;
  message: string;
  emailSent: boolean;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;