# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Event and organizer operations
  - name: check-in
    description: Attendee self-check-in operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /events:
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
  /events/{eventId}/import:
    post:
      operationId: importAttendees
      tags: [events]
      summary: Import attendee and coupon rows
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendeeImportInput"
      responses:
        "200":
          description: Attendees imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /events/{eventId}/summary:
    get:
      operationId: getEventSummary
      tags: [events]
      summary: Get check-in summary
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event check-in summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /events/{eventId}/attendees:
    get:
      operationId: listAttendees
      tags: [events]
      summary: List attendees without coupon codes
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Attendee statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendeeStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /events/{eventId}/final-report:
    post:
      operationId: sendFinalReport
      tags: [events]
      summary: Email the organizer the final attendance report
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Final report sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalReportResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /events/{eventId}/check-in:
    post:
      operationId: checkInAttendee
      tags: [check-in]
      summary: Check in an attendee by registered email
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "200":
          description: Attendee checked in or already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Event:
      type: object
      required:
        - id
        - name
        - eventDate
        - dataDeleteDate
        - attendeeCount
        - checkedInCount
        - checkInUrl
      properties:
        id:
          type: string
        name:
          type: string
        eventDate:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        dataDeleteDate:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        attendeeCount:
          type: integer
        checkedInCount:
          type: integer
        checkInUrl:
          type: string
          format: uri
    EventInput:
      type: object
      required:
        - name
        - organizerEmail
        - eventDate
        - dataDeleteDate
      properties:
        name:
          type: string
          minLength: 1
        organizerEmail:
          type: string
          format: email
        eventDate:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        dataDeleteDate:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
    AttendeeImportInput:
      type: object
      required:
        - attendees
      properties:
        attendees:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AttendeeImportRow"
    AttendeeImportRow:
      type: object
      required:
        - email
        - couponCode
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        couponCode:
          type: string
          minLength: 1
    ImportResult:
      type: object
      required:
        - imported
        - rejected
      properties:
        imported:
          type: integer
        rejected:
          type: integer
        errors:
          type: array
          items:
            type: string
    EventSummary:
      type: object
      required:
        - total
        - checkedIn
        - remaining
        - emailSent
      properties:
        total:
          type: integer
        checkedIn:
          type: integer
        remaining:
          type: integer
        emailSent:
          type: integer
    FinalReportResult:
      type: object
      required:
        - sent
        - total
        - checkedIn
        - didNotShow
      properties:
        sent:
          type: boolean
        total:
          type: integer
        checkedIn:
          type: integer
        didNotShow:
          type: integer
    AttendeeStatus:
      type: object
      required:
        - email
        - checkedIn
        - emailSent
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        checkedIn:
          type: boolean
        checkedInAt:
          type: ["string", "null"]
          format: date-time
        emailSent:
          type: boolean
    CheckInInput:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    CheckInResult:
      type: object
      required:
        - status
        - message
        - emailSent
      properties:
        status:
          type: string
          enum: [checked_in, already_checked_in]
        message:
          type: string
        emailSent:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string