# lib/api-zod/src/generated/types/attendeeImportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendeeImportRow } from './attendeeImportRow';

export interface AttendeeImportInput {
  /** @minItems 1 */
  attendees: AttendeeImportRow[];
}