# lib/api-zod/src/generated/types/attendeeImportRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttendeeImportRow {
  name?: string;
  email: string;
  /** @minLength 1 */
  couponCode: string;
}