# lib/api-zod/src/generated/types/attendeeStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttendeeStatus {
  name?: string;
  email: string;
  checkedIn: boolean;
  /** @nullable */
  checkedInAt?: Date | null;
  emailSent: boolean;
}