# lib/api-zod/src/generated/types/checkInResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckInResultStatus } from './checkInResultStatus';

export interface CheckInResult {
  status: CheckInResultStatus;
  message: string;
  emailSent: boolean;
}