# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: string;
  name: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  eventDate: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  dataDeleteDate: string;
  attendeeCount: number;
  checkedInCount: number;
  checkInUrl: string;
}