# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  /** @minLength 1 */
  name: string;
  organizerEmail: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  eventDate: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  dataDeleteDate: string;
}