# lib/api-zod/src/generated/types/eventSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EventSummary {
  total: number;
  checkedIn: number;
  remaining: number;
  emailSent: number;
}