# lib/api-zod/src/generated/types/finalReportResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FinalReportResult {
  sent: boolean;
  total: number;
  checkedIn: number;
  didNotShow: number;
}