# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImportResult {
  imported: number;
  rejected: number;
  errors?: string[];
}